Started Notification Manager
Imported the pods and stated the Notification manager class. The banner seems to work well. The popup needs to be done (I would like to work with someone in design to know how the popup should look)
I am currently waiting for some changed to be pulled from the error class before I do testing with this class.