Implemented the chat details button (#122)
* Implemented the chat details button

The chat details button currently only holds test data.
I made the assumption that the ChatScreenVC would be able to pass the necessary information to the ChatDetailsVC, making a network request unnecessary.

A special thanks to @hirdaygupta for his patience.

* Populate pop up fields correctly

* Add back commented test data

So that @kamalca can easily debug

* Improve utility func scoping

* Made the popup less tall

Popup was arbitrarily tall before, now it is only big enough to fit all the objects. Real nice and flexible now. Like a yoga instructor.