Chat list profile picture
I added code that loads the images in the `ChatListVC`, but I was unable to test the code yet. I will do testing once I learn how to add a match to my user.

I increased the compression on the images because they were taking too long to load. I'm prepared to defend my decision.